tests/wpt/web-platform-tests/web-bundle/subresource-loading/subresource-loading-path-restriction.https.tentative.html

<!DOCTYPE html>
<title>Path restriction on subresource loading with WebBundles</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/master/explainers/subresource-loading.md"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <link
    rel="webbundle"
    href="../resources/wbn/path-restriction.wbn"
    resources="https://web-platform.test:8444/web-bundle/resources/wbn/resource.js
                   https://web-platform.test:8444/web-bundle/resources/wbn/sub/resource.js
                   https://web-platform.test:8444/web-bundle/resources/wbn-resource.js
                   https://web-platform.test:8444/web-bundle/resources/wbn1/resource.js
                   https://web-platform.test:8444/web-bundle/resources/other/resource.js
                   https://web-platform.test:8444/web-bundle/resources/resource.js"
  />
  <script>
    promise_test(async () => {
      const resources = [
        "https://web-platform.test:8444/web-bundle/resources/wbn/resource.js",
        "https://web-platform.test:8444/web-bundle/resources/wbn/sub/resource.js",
      ];
      for (const resource of resources) {
        const response = await fetch(resource);
        assert_true(response.ok, resource + " should be loaded");
      }
    }, "Subresources should be loaded.");

    promise_test(async () => {
      const resources = [
        "https://web-platform.test:8444/web-bundle/resources/wbn-resource.js",
        "https://web-platform.test:8444/web-bundle/resources/wbn1/resource.js",
        "https://web-platform.test:8444/web-bundle/resources/other/resource.js",
        "https://web-platform.test:8444/web-bundle/resources/resource.js",
      ];
      for (const resource of resources) {
        const response = await fetch(resource);
        assert_false(response.ok, resource + " should not be loaded");
      }
    }, "Subresources should not be loaded due to path restriction.");
  </script>
</body>
